+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to the "xycode" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.6]
 
-- Initial release
+-   `xycode.openconfig` command: open config directory.
+-   encoding config option.
+
+## [0.0.5]
+
+-   sfdx metadata diff
+
+## [0.0.4]
+
+-   ConfigManager: selectable config
+-   fix notShowProcess bug
+
+## [0.0.3]
+
+-   fix config directory not found error
+
+## [0.0.2]
+
+-   Initial Document
+
+## [0.0.1]
+
+-   Initial release